requirements.txt
langchain
langgraph
langchain-community
chromadb
sentence-transformers
python-dotenv
openai
pypdf
torch
transformers
accelerate
